backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
openai==0.28.1
requests==2.31.0
pydantic==2.5.2
python-dotenv==1.0.0

